requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
greenlet==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
